Correction stats europeebes

+++ project/plugins/acVinDRMPlugin/lib/DRMImportCsvEdi.class.php
@@ -685,7 +685,7 @@ class DRMImportCsvEdi extends DRMCsvEdi {
                         break;
                     case self::TYPE_ANNEXE_STATS_EUROPEENES:
                         $typeStat = strtolower(KeyInflector::slugify($csvRow[self::CSV_ANNEXE_TYPEMVT]));
-                        $stat = str_replace(',', '.', $csvRow[self::CSV_ANNEXE_QUANTITE]);
+                        $stat = (float) str_replace(',', '.', $csvRow[self::CSV_ANNEXE_QUANTITE]);
                         if(!in_array($typeStat, array('jus', 'mcr', 'vinaigre'))) {
                             if ($just_check) {
                                 $this->csvDoc->addErreur($this->annexesTypeMvtStatistiquesWrongFormatError($num_ligne, $csvRow));
